requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-annotate
#
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
django==3.1.3
djangorestframework-camel-case==1.2.0
djangorestframework==3.12.2
drf-registration==0.1.0
idna==2.10
python-decouple==3.3
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
